fix: remove unused utils files
